# [bernoulli_kernel] Replace CPU_tensor_apply functions with cpu_serial_kernel