@RockingJavaBean 2020-06-11
The `TensorIterator` is constructed for the `cpu_serial_kernel` function to replace CPU_tensor_apply functions.
Please kindly review this PR. @pbelevich