@RockingJavaBean 2020-06-12
@pbelevich Thank you so much for the kind review and providing the performance test script.

I have tested the performance using ipython's "timeit" on my laptop (`Intel Core i5-1035G1`).
Below is the test result.

```python
import torch
size = 1000000

t = torch.empty(size, dtype=torch.int32, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')

timeit t.bernoulli_(p)
------------------
Before: 11.4 ms ± 61.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 14.5 ms ± 686 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.int64, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')

timeit t.bernoulli_(p)
------------------
Before: 11.6 ms ± 336 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 14.2 ms ± 255 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.float32, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')

timeit t.bernoulli_(p)
------------------
Before: 11.1 ms ± 73.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 9.58 ms ± 82 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.float64, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')

timeit t.bernoulli_(p)
------------------
Before: 11.7 ms ± 447 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 13.1 ms ± 392 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.bool, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')

timeit t.bernoulli_(p)
------------------
Before: 21.9 ms ± 247 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
After: 25.8 ms ± 322 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
------------------

t = torch.empty(size, dtype=torch.int64, device='cpu')
p = torch.rand(size, dtype=torch.float64, device='cpu')

timeit t.bernoulli_(p)
------------------
Before: 15 ms ± 159 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: RuntimeError: !needs_dynamic_casting<func_t>::check(iter) INTERNAL ASSERT FAILED at "..\\aten\\src\\ATen/native/cpu/Loops.h":245, please report a bug to PyTorch.
------------------

t = torch.empty(size, dtype=torch.float32, device='cpu')
p = torch.rand(size, dtype=torch.float64, device='cpu')

timeit t.bernoulli_(p)
------------------
Before: 15.1 ms ± 79.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: RuntimeError: !needs_dynamic_casting<func_t>::check(iter) INTERNAL ASSERT FAILED at "..\\aten\\src\\ATen/native/cpu/Loops.h":245, please report a bug to PyTorch.
------------------

t = torch.empty(size, dtype=torch.float64, device='cpu')
p = torch.rand(size, dtype=torch.float64, device='cpu')

timeit t.bernoulli_(p)
-------------------
Before: 15.2 ms ± 112 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 14 ms ± 315 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
-------------------

t = torch.empty(size, dtype=torch.bool, device='cpu')
p = torch.rand(size, dtype=torch.float64, device='cpu')

timeit t.bernoulli_(p)
-------------------
Before: 24.9 ms ± 297 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
After: RuntimeError: !needs_dynamic_casting<func_t>::check(iter) INTERNAL ASSERT FAILED at "..\\aten\\src\\ATen/native/cpu/Loops.h":245, please report a bug to PyTorch.
-------------------
```

Compared with ` CPU_tensor_apply2`, using `TensorIterator` with `cpu_serial_kernel` is

- **slower**: when the dtype of input `tensor p` is `float32`, and the dytpe of `tensor t ` is not the `float32`.
- **similar performance**: when `tensor p` and `tensor t` have the same dtype (float32/float64). The performance is even better on my laptop.
- **runtime error**: when the dtype of input `tensor p` is `float64`, and `t` is not a `float64` tensor.

The runtime error is thrown by `aten/src/ATen/native/cpu/Loops.h` due to the failure of the `TORCH_INTERNAL_ASSERT`.
https://github.com/pytorch/pytorch/blob/246d7bb41d138a3b05834c17311f2c326d1d2536/aten/src/ATen/native/cpu/Loops.h#L244-L245

The possible causes for the failed assert in the `loops.h` when input tensor is `float64` are needed to be found out.
My guess is the failed assert is due to the new behaviors and checks of `TensorIterator`.