@RockingJavaBean 2020-06-13
The performance degradation of `TensorIterator with cpu_serial_kernel` will have to do with the casting behaviors of `TensorIterator` when the input and output tensor do not share the same dtype.