@RockingJavaBean 2020-06-15
@pbelevich thanks so much for your kind help and guidance through this pull request.
I have updated the code according to the feedback, and enhance the test method `test_bernoulli_self` by adding more float dtypes to the input tensor p.

I also manually re-tested the performance using `ipython's timeit` on my laptop.
Below is the test result, while `after` is the performance of latest PR.

```python
import torch
size = 1000000

t = torch.empty(size, dtype=torch.int32, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')
timeit t.bernoulli_(p)
------------------
Before: 11.4 ms ± 61.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 9.77 ms ± 85.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.int64, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')
timeit t.bernoulli_(p)
------------------
Before: 11.6 ms ± 336 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 10.5 ms ± 147 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.float32, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')
timeit t.bernoulli_(p)
------------------
Before: 11.1 ms ± 73.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 9.5 ms ± 108 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.float64, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')
timeit t.bernoulli_(p)
------------------
Before: 11.7 ms ± 447 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 9.82 ms ± 111 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.bool, device='cpu')
p = torch.rand(size, dtype=torch.float32, device='cpu')
timeit t.bernoulli_(p)
------------------
Before: 21.9 ms ± 247 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
After: 20.9 ms ± 148 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
------------------

t = torch.empty(size, dtype=torch.int64, device='cpu')
p = torch.rand(size, dtype=torch.float64, device='cpu')
timeit t.bernoulli_(p)
------------------
Before: 15 ms ± 159 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 13.2 ms ± 56.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.float32, device='cpu')
p = torch.rand(size, dtype=torch.float64, device='cpu')
timeit t.bernoulli_(p)
------------------
Before: 15.1 ms ± 79.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 13.9 ms ± 274 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
------------------

t = torch.empty(size, dtype=torch.float64, device='cpu')
p = torch.rand(size, dtype=torch.float64, device='cpu')
timeit t.bernoulli_(p)
-------------------
Before: 15.2 ms ± 112 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
After: 14 ms ± 176 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
-------------------

t = torch.empty(size, dtype=torch.bool, device='cpu')
p = torch.rand(size, dtype=torch.float64, device='cpu')
timeit t.bernoulli_(p)
-------------------
Before: 24.9 ms ± 297 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
After: 24 ms ± 512 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
---------------
```

No performance degradation and runtime errors are found during the manual test.